+++ src/resources/assets/js/components/annotationShapes/circle.vue
