+++ src/resources/assets/js/components/annotationShapes/ellipse.vue
