+++ src/resources/assets/js/components/annotationShapes/line.vue
