+++ src/resources/assets/js/components/annotationShapes/mixins/shape.vue
