+++ src/resources/assets/js/components/annotationShapes/polygon.vue
