+++ src/resources/assets/js/components/annotationShapes/rectangle.vue
